- Now using Pyiceberg standard configuration means - Using DuckDB to process the SELECT query on a full PyIceberg scan
